+++ dataprovider/foxsports/runner/eventdata/scraper_nba_box_score.go
@@ -27,7 +27,9 @@ func (s *NBABoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
 	matchupModel := matchup.(model.Matchup)
 	var context Context
 	context.AwayTeam = matchupModel.AwayTeamNameFull
+	context.AwayID = matchupModel.AwayTeamID
 	context.HomeTeam = matchupModel.HomeTeamNameFull
+	context.HomeID = matchupModel.HomeTeamID
 	context.EventID = matchupModel.EventID
 	var data []interface{}
 	// Construct event data URL
@@ -38,12 +40,14 @@ func (s *NBABoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
 		return OutputWrapper{Error: err, Context: context}
 	}
 	context.URL = url
+	pullTimestamp := time.Now().UTC()
 	// Fetch event data
 	responseBody, err := s.FetchEventData(url)
 	if err != nil {
 		log.Println("Issue fetching event data")
 		return OutputWrapper{Error: err, Context: context}
 	}
+	context.PullTimestamp = pullTimestamp
 	// Unmarshal JSON
 	var responsePayload jsonresponse.NBAEventData
 	err = json.Unmarshal(responseBody, &responsePayload)
@@ -175,22 +179,471 @@ func (s *NBABoxScoreScraper) Scrape(matchup interface{}) OutputWrapper {
 	}
 
 	// parse and develop playerMap of relevant statlines
-	playerMap, err := s.parseBoxScoreStats(responsePayload)
+	playerMap, err := s.parseBoxScoreStats(responsePayload, context)
 	if err != nil {
 		return OutputWrapper{Error: err, Context: context}
 	}
 
 	// Allocate each statline to data output
 	for _, obj := range playerMap {
-		data = append(data, obj)
+		data = append(data, *obj)
 	}
 	diff := time.Now().UTC().Sub(start)
 	log.Printf("Scraping of event %d (%s vs %s) completed in %s\n", matchupModel.EventID, matchupModel.AwayTeamNameFull, matchupModel.HomeTeamNameFull, diff)
 	return OutputWrapper{Output: data, Context: context}
 }
 
-func (s *NBABoxScoreScraper) parseBoxScoreStats(responsePayload jsonresponse.NBAEventData) (map[int64]model.NBABoxScoreStats, error) {
-	var playerMap map[int64]model.NBABoxScoreStats
+func (s *NBABoxScoreScraper) parseBoxScoreStats(responsePayload jsonresponse.NBAEventData, context Context) (map[int64]*model.NBABoxScoreStats, error) {
+	// var playerMap map[int64]*model.NBABoxScoreStats
+	playerMap := make(map[int64]*model.NBABoxScoreStats)
+
+	// Home
+	// Starters
+	starterRecords := responsePayload.BoxScore.BoxScoreSections.HomePlayerStats.BoxscoreItems[0].BoxscoreTable.Rows
+	for _, record := range starterRecords {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+
+		playerMap[playerID] = &model.NBABoxScoreStats{
+			PullTimestamp: context.PullTimestamp,
+			PlayerID:      playerID,
+			EventID:       context.EventID,
+			Starter:       true,
+			TeamID:        context.HomeID,
+			Team:          context.HomeTeam,
+			OpponentID:    context.AwayID,
+			Opponent:      context.AwayTeam,
+		}
+		playerMap[playerID].Player = record.EntityLink.Player
+		playerMap[playerID].Position = *record.Columns[0].Superscript
+		playerMap[playerID].MinutesPlayed, err = util.TextToInt(record.Columns[1].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].OffensiveRebounds, err = util.TextToInt(record.Columns[2].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].DefensiveRebounds, err = util.TextToInt(record.Columns[3].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].TotalRebounds, err = util.TextToInt(record.Columns[4].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Assists, err = util.TextToInt(record.Columns[5].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Steals, err = util.TextToInt(record.Columns[6].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Blocks, err = util.TextToInt(record.Columns[7].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Turnovers, err = util.TextToInt(record.Columns[8].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].PersonalFouls, err = util.TextToInt(record.Columns[9].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Points, err = util.TextToInt(record.Columns[10].Text)
+		if err != nil {
+			return playerMap, err
+		}
+	}
+
+	// Bench
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.HomePlayerStats.BoxscoreItems[1].BoxscoreTable.Rows {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+
+		playerMap[playerID] = &model.NBABoxScoreStats{
+			PullTimestamp: context.PullTimestamp,
+			PlayerID:      playerID,
+			EventID:       context.EventID,
+			TeamID:        context.HomeID,
+			Team:          context.HomeTeam,
+			OpponentID:    context.AwayID,
+			Opponent:      context.AwayTeam,
+		}
+		playerMap[playerID].Player = record.EntityLink.Player
+		playerMap[playerID].Position = *record.Columns[0].Superscript
+		// MinutesPlayed
+		if record.Columns[1].Text == "-" {
+			playerMap[playerID].MinutesPlayed = 0
+		} else {
+			playerMap[playerID].MinutesPlayed, err = util.TextToInt(record.Columns[1].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// OffensiveRebounds
+		if record.Columns[2].Text == "-" {
+			playerMap[playerID].OffensiveRebounds = 0
+		} else {
+			playerMap[playerID].OffensiveRebounds, err = util.TextToInt(record.Columns[2].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// DefensiveRebounds
+		if record.Columns[3].Text == "-" {
+			playerMap[playerID].DefensiveRebounds = 0
+		} else {
+			playerMap[playerID].DefensiveRebounds, err = util.TextToInt(record.Columns[3].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// TotalRebounds
+		if record.Columns[4].Text == "-" {
+			playerMap[playerID].TotalRebounds = 0
+		} else {
+			playerMap[playerID].TotalRebounds, err = util.TextToInt(record.Columns[4].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Assists
+		if record.Columns[5].Text == "-" {
+			playerMap[playerID].Assists = 0
+		} else {
+			playerMap[playerID].Assists, err = util.TextToInt(record.Columns[5].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Steals
+		if record.Columns[6].Text == "-" {
+			playerMap[playerID].Steals = 0
+		} else {
+			playerMap[playerID].Steals, err = util.TextToInt(record.Columns[6].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Blocks
+		if record.Columns[7].Text == "-" {
+			playerMap[playerID].Blocks = 0
+		} else {
+			playerMap[playerID].Blocks, err = util.TextToInt(record.Columns[7].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Turnovers
+		if record.Columns[8].Text == "-" {
+			playerMap[playerID].Turnovers = 0
+		} else {
+			playerMap[playerID].Turnovers, err = util.TextToInt(record.Columns[8].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// PersonalFouls
+		if record.Columns[9].Text == "-" {
+			playerMap[playerID].PersonalFouls = 0
+		} else {
+			playerMap[playerID].PersonalFouls, err = util.TextToInt(record.Columns[9].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Points
+		if record.Columns[10].Text == "-" {
+			playerMap[playerID].Points = 0
+		} else {
+			playerMap[playerID].Points, err = util.TextToInt(record.Columns[10].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+	}
+
+	// Shooting
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.HomePlayerStats.BoxscoreItems[3].BoxscoreTable.Rows {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+		_, exists := playerMap[playerID]
+		if !exists {
+			return playerMap, fmt.Errorf("Shooting stats are unavailable for player id, %d, on team, %s. This is unexpected and should not be a realistic scenario.", playerID, context.HomeTeam)
+		}
+		// Field goals
+		fgSplit := strings.Split(record.Columns[1].Text, "-")
+		playerMap[playerID].FieldGoalsMade, err = util.TextToInt(fgSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].FieldGoalAttempts, err = util.TextToInt(fgSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+		// Threes
+		threesSplit := strings.Split(record.Columns[2].Text, "-")
+		playerMap[playerID].ThreePointsMade, err = util.TextToInt(threesSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].ThreePointAttempts, err = util.TextToInt(threesSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+		// Freethrows
+		freeThrowsSplit := strings.Split(record.Columns[3].Text, "-")
+		playerMap[playerID].FreeThrowsMade, err = util.TextToInt(freeThrowsSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].FreeThrowAttempts, err = util.TextToInt(freeThrowsSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+	}
+
+	// Away
+	// Starters
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.AwayPlayerStats.BoxscoreItems[0].BoxscoreTable.Rows {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+
+		playerMap[playerID] = &model.NBABoxScoreStats{
+			PullTimestamp: context.PullTimestamp,
+			PlayerID:      playerID,
+			EventID:       context.EventID,
+			Starter:       true,
+			OpponentID:    context.HomeID,
+			Opponent:      context.HomeTeam,
+			TeamID:        context.AwayID,
+			Team:          context.AwayTeam,
+		}
+		playerMap[playerID].Player = record.EntityLink.Player
+		playerMap[playerID].Position = *record.Columns[0].Superscript
+		playerMap[playerID].MinutesPlayed, err = util.TextToInt(record.Columns[1].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].OffensiveRebounds, err = util.TextToInt(record.Columns[2].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].DefensiveRebounds, err = util.TextToInt(record.Columns[3].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].TotalRebounds, err = util.TextToInt(record.Columns[4].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Assists, err = util.TextToInt(record.Columns[5].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Steals, err = util.TextToInt(record.Columns[6].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Blocks, err = util.TextToInt(record.Columns[7].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Turnovers, err = util.TextToInt(record.Columns[8].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].PersonalFouls, err = util.TextToInt(record.Columns[9].Text)
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].Points, err = util.TextToInt(record.Columns[10].Text)
+		if err != nil {
+			return playerMap, err
+		}
+	}
+
+	// Bench
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.AwayPlayerStats.BoxscoreItems[1].BoxscoreTable.Rows {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+
+		playerMap[playerID] = &model.NBABoxScoreStats{
+			PullTimestamp: context.PullTimestamp,
+			PlayerID:      playerID,
+			EventID:       context.EventID,
+			OpponentID:    context.HomeID,
+			Opponent:      context.HomeTeam,
+			TeamID:        context.AwayID,
+			Team:          context.AwayTeam,
+		}
+		playerMap[playerID].Player = record.EntityLink.Player
+		playerMap[playerID].Position = *record.Columns[0].Superscript
+		// MinutesPlayed
+		if record.Columns[1].Text == "-" {
+			playerMap[playerID].MinutesPlayed = 0
+		} else {
+			playerMap[playerID].MinutesPlayed, err = util.TextToInt(record.Columns[1].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// OffensiveRebounds
+		if record.Columns[2].Text == "-" {
+			playerMap[playerID].OffensiveRebounds = 0
+		} else {
+			playerMap[playerID].OffensiveRebounds, err = util.TextToInt(record.Columns[2].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// DefensiveRebounds
+		if record.Columns[3].Text == "-" {
+			playerMap[playerID].DefensiveRebounds = 0
+		} else {
+			playerMap[playerID].DefensiveRebounds, err = util.TextToInt(record.Columns[3].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// TotalRebounds
+		if record.Columns[4].Text == "-" {
+			playerMap[playerID].TotalRebounds = 0
+		} else {
+			playerMap[playerID].TotalRebounds, err = util.TextToInt(record.Columns[4].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Assists
+		if record.Columns[5].Text == "-" {
+			playerMap[playerID].Assists = 0
+		} else {
+			playerMap[playerID].Assists, err = util.TextToInt(record.Columns[5].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Steals
+		if record.Columns[6].Text == "-" {
+			playerMap[playerID].Steals = 0
+		} else {
+			playerMap[playerID].Steals, err = util.TextToInt(record.Columns[6].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Blocks
+		if record.Columns[7].Text == "-" {
+			playerMap[playerID].Blocks = 0
+		} else {
+			playerMap[playerID].Blocks, err = util.TextToInt(record.Columns[7].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Turnovers
+		if record.Columns[8].Text == "-" {
+			playerMap[playerID].Turnovers = 0
+		} else {
+			playerMap[playerID].Turnovers, err = util.TextToInt(record.Columns[8].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// PersonalFouls
+		if record.Columns[9].Text == "-" {
+			playerMap[playerID].PersonalFouls = 0
+		} else {
+			playerMap[playerID].PersonalFouls, err = util.TextToInt(record.Columns[9].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+		// Points
+		if record.Columns[10].Text == "-" {
+			playerMap[playerID].Points = 0
+		} else {
+			playerMap[playerID].Points, err = util.TextToInt(record.Columns[10].Text)
+			if err != nil {
+				return playerMap, err
+			}
+		}
+	}
+
+	// Shooting
+	for _, record := range responsePayload.BoxScore.BoxScoreSections.AwayPlayerStats.BoxscoreItems[3].BoxscoreTable.Rows {
+		if record.EntityLink == nil {
+			continue
+		}
+		playerID, err := util.TextToInt64(record.EntityLink.Layout.Tokens.ID)
+		if err != nil {
+			return playerMap, err
+		}
+		_, exists := playerMap[playerID]
+		if !exists {
+			return playerMap, fmt.Errorf("Shooting stats are unavailable for player id, %d, on team, %s. This is unexpected and should not be a realistic scenario.", playerID, context.AwayTeam)
+		}
+		// Field goals
+		fgSplit := strings.Split(record.Columns[1].Text, "-")
+		playerMap[playerID].FieldGoalsMade, err = util.TextToInt(fgSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].FieldGoalAttempts, err = util.TextToInt(fgSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+		// Threes
+		threesSplit := strings.Split(record.Columns[2].Text, "-")
+		playerMap[playerID].ThreePointsMade, err = util.TextToInt(threesSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].ThreePointAttempts, err = util.TextToInt(threesSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+		// Freethrows
+		freeThrowsSplit := strings.Split(record.Columns[3].Text, "-")
+		playerMap[playerID].FreeThrowsMade, err = util.TextToInt(freeThrowsSplit[0])
+		if err != nil {
+			return playerMap, err
+		}
+		playerMap[playerID].FreeThrowAttempts, err = util.TextToInt(freeThrowsSplit[1])
+		if err != nil {
+			return playerMap, err
+		}
+	}
+
 	return playerMap, nil
 
 }